Solucionado fallo casilla bloqueada
Eliminados los restos de la secuencia de victoria antigua, creada una nueva secuencia de victoria.
