+++ src/store/index.js
@@ -4,14 +4,14 @@ import _ from 'lodash'
 
 
 const tableroInicial = [
-  [2, 2, 2, 2, 2, 2, 2, 0],
-  [2, 2, 1, 1, 1, 1, 1, 1],
-  [2, 1, 1, 1, 1, 1, 1, 1],
-  [2, 1, 1, 1, 1, 1, 1, 1],
-  [2, 1, 1, 1, 1, 1, 1, 1],
-  [1, 1, 1, 1, 1, 1, 1, 1],
-  [1, 1, 1, 1, 1, 1, 1, 1],
-  [1, 1, 1, 1, 1, 1, 1, 1]
+  [0, 0, 0, 0, 0, 0, 0, 0],
+  [0, 0, 0, 0, 0, 0, 0, 0],
+  [0, 0, 0, 0, 0, 0, 0, 0],
+  [0, 0, 0, 1, 2, 0, 0, 0],
+  [0, 0, 0, 2, 1, 0, 0, 0],
+  [0, 0, 0, 0, 0, 0, 0, 0],
+  [0, 0, 0, 0, 0, 0, 0, 0],
+  [0, 0, 0, 0, 0, 0, 0, 0]
 ];
 
 Vue.use(Vuex)
@@ -46,6 +46,8 @@ export default new Vuex.Store({
 
     contadorFichasNegras: 2,
     contadorFichasBlancas: 2,
+
+    posiciones: 0
 
   },
   mutations: {
